requirement testing

+++ 0x01-python_async_function/0-basic_async_syntax.py
@@ -8,15 +8,14 @@ import asyncio
 import random
 
 
-# Define an asynchronous coroutine called wait_random
-async def wait_random(max_delay: float = 10) -> float:
+async def wait_random(max_delay: int = 10) -> float:
     """
     Asynchronous coroutine that waits for a random delay between 0 and
     max_delay seconds.
 
     Args:
-        max_delay (float, optional): The maximum delay in seconds.
-        Defaults to 10.
+        max_delay (int, optional): The maximum delay in seconds. Defaults
+        to 10.
 
     Returns:
         float: The actual delay that was used.
